Refactored headers => only one header (#29)
