+++ src/c/types/array.c
@@ -12,8 +12,6 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-#include <ucdr/types/array.h>
-#include <ucdr/types/basic.h>
 #include "../common_internal.h"
 
 #include <string.h>
