+++ src/c/types/string.c
@@ -13,8 +13,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-#include <ucdr/types/string.h>
-#include <ucdr/types/sequence.h>
+#include <ucdr/microcdr.h>
 
 #include <string.h>
